Split CircuitDef trait
The CircuitDef is a trait that is implemented by rust code generated from the
circuit for both the verifier and the prover. Since we need to split off the
recursive verifier, we need to split this trait into two parts: one that is
used in both the prover and the verifier and one that is used only in the
prover. By doing so, we can keep the prover private for the timebeing. This
change implements the split and adds a CircuitProveImpl in addition to the
existing CircuitImpl.