Publish Recursion Verifier (#603)
This change moves the recursion verifier from the private repository. This is
done by splitting the `CircuitDef` trait and adding files used by the recursion
verifier.

The CircuitDef is a trait that is implemented by rust code generated from the
circuit for both the verifier and the prover. In order to split off the recursive
verifier, we need to split this trait into two parts: one that is used in both the
prover and the verifier and one that is used only in the prover. The former trait
called `CircuitCoreDef` is implemented for the recursive verifier which requires
generated rust files from the recursion circuit.