implement SessionReceipt trait (#616)
This pull request implements the `SessionReceipt` trait that is implemented by
`SessionRollupReceipt` and `SessionFlatReceipt` (also known as the old
`SessionReceipt`). This is useful because it allows the host code to call verify
and to get the journal contents regardless of the concrete type of session
receipt. This trait is practical when a developer has a code base that uses
bonsai resulting in a `SessionRollupReceipt` or from a local prover resulting in a
`SessionFlatReceipt`.  Examples have been updated to use this new trait. The main
difference is to replace existing `receipt.journal` with `receipt.get_journal()`
and `receipt.verify(XXXX_ID)` with `receipt.verify(XXXX_ID.into())`.