Emulator: Improve usability by restricting memory region access for guest programs (#727)
Memory regions outside of the address range `TEXT_START` to `SYSTEM.start() -
1` are used to store information related to the zkVM. Tampering with this
region will cause the prover to crash. Restrict the guest from writing to this
region.

* Executor: fix panic by guarding against invalid page indicies

It was possible for the executor to panic when accessing higher regions of
memory. This change fixes the panics by checking that the address trying to be
accessed has a valid page index.