Add new verify overload to IRiscZeroVerifier.sol (#782)
Add an overload for the verify function that takes the full journal data instead of only the journal bytes. This provides another methods for calling the verify function, which in this case does not assume the caller has pre-hashed the data.

This also allows an on-chain verify to be called fully as a view call, or as a standalone transaction which is useful for testing, development, and demos.