Add entropy to memory state to protect privacy (#812)
A Merkle root of the memory state is included in the `ReceiptMetadata`,
and so is made public to the verifier. Without adding entropy, this
leaks a commitment to the end memory state that may reveal information
about the program state including inputs.

This PR adds 128 bits of entropy, provided by the host, to the memory
state on guest initialization to prevent this privacy leakage.