Expose fault_monitor to those that actually need it
Prior to this change, I was exposing the fault_monitor to all guest that have
std enabled. This led to the use of an unidiomatic use of `#[cfg(feature = "std")]`.
This was done because fault_monitor uses rrs-lib which depends on std. However,
the only use for fault monitor are in the prover and in the fault checker
guest. Create yet another feature flag called fault-proof and expose
fault_monitor conditionally on this flag rather than "std".