load_elf: change maximum memory address to SYSTEM.start()
load_elf takes a max_mem parameter before this change, MEM_SIZE was used as the
parameter. Any memory region about `SYSTEM.start()` is intended for use by the
zkVM and is not to be used by the ELF. Rather than have maximum memory as a
parameter, use a hard-coded constant, GUEST_MAX_MEM as the upper bound for
memory during elf load.