load_elf: use GUEST_MAX_MEM instead of MEM_SIZE and do not re-export MEM_SIZE in risc0-zkvm
MEM_SIZE indicates the size of the total memory in the zkVM. This information
is not very useful because what users are actually concerned about are the
maximum memory that guest programs can use. As of this writing anything between
0x400 and SYSTEM.start can be used for the guest. Anything between SYSTEM.start
to MEM_SIZE should not be used by guests or it could result in unintended
behavior in the guest program. Users call the load_elf function to create an
instances of valid `Program` structs. Using MEM_SIZE as a parameter to this
function allows load_elf to write between invalid regions for guest programs
but it also enables this function to write between SYSTEM.start and MEM_SIZE
which results in undesirable behavior. Prevent this by passing GUEST_MAX_MEM
instead of MEM_SIZE to load_elf.

Thinking about the usecase of MEM_SIZE, I didn't find that this is a useful
value for the zkVM to export becasue users are actually interested in
GUEST_MAX_MEM rather than MEM_SIZE. I've removed the re-export for MEM_SIZE
from risc0-zkvm.