reproducible builds: move code to `risc0-build`
The reproducible build implementation used to be a part of the cargo-risczero
utility. This change moves the code from cargo-risczero to risc0-build. By
doing so, we will be able to integrate the docker builds as a part of the
risc0-build mechanism. Eventually, it would be nice for users to be able to
build guest code using by setting feature flags for the risc0-build crate.

This code movement is the first step in facilitating users to use docker to
build their guest code.