tests: place some tests under test-exact-cycles feature flag
The tests being gated by this feature flag measure cycles and segments and are
extremely sensitive to changes in the elf binary. We have seen cases where CI
machines fail these tests on different architectures depending on the commit.
What's interesting is that the test behavior changes even for commits that did
not actually change the test. The root of this problem lies in the fact that
rust does not support reproducible builds so that each architecture is running
slightly mismatched ELF binaries from eachother.

This is an attemp to eliminate test failures that happen from reproducible
builds. The tests under this new flag must be run on elfs generated by the
docker environment. All others can be run by the usual `cargo test` command.