Add experimental cargo-risczero test command and support for building rust-runtime.a (#646)
This PR adds a new `cargo riszero test` command, which can be used to
build the current crate for
the RISC Zero target, and run it's tests inside the Executor.

This is intended to allow developers of libraries intended to support
the RISC Zero zkVM a way to
test their code, using all existing tests and any new ones they write
for zkVM specific
functionality, without writing a guest wrapper or duplicating effort. As
an example, see
https://github.com/risc0/RustCrypto-crypto-bigint/pull/2 which adds
support for running all tests
for the crypto-bigint library, which has been modified to use the bigint
multiplication accelerator
when built for the RISC Zero target. This may also be used to test crate
compatibility with the zkVM
without any modifications e.g. in the crate validator.

As a dependency, this PR includes support for building the zkVM rust
runtime as a linkable `.a`
file. This is used within the `cargo risczero test` command to link
required functionality, such as
the syscalls, for the test binaries. More generally it can be used as a
tool for developing other
runtimes, as the symbols can be used in any language with C FFI.

---------

Co-authored-by: nils <shkoo@users.noreply.github.com>
Co-authored-by: Joby Thundil <joby@risczero.com>