zkVM tests: add feature flag to build multi-test with docker environment (#912)
This PR adds a test-exact-cycles feature for risc0-zkvm. This is used to
build test guest binaries like multi-test using the docker environment
in the `cargo test` command. The intention here is to run the tests
using a reproducible ELF binary to eliminate test failures across
different architectures that are caused by entropy in the ELF binaries
resulting from rust build tools.

I've gated 4 tests using this feature flag. My solution is to create a new
feature flag that the CI can use and it will run all tests including the ones
that rely on exact cycles counts. If users do not wish to run the tests
that require the reproducible binary, they can simply leave out the
`test-exact-cycles` feature flag in their `cargo test` invocation.

Changes in this PR involves moving code from the `cargo risczero build`
command to the `risc0-build` crate, and making the functionality similar to
the existing embed_methods() function. 

---------

Co-authored-by: Frank Laub <flaub@risc0.com>