guest memory: place stack below text/code
The previous memory layout was prone to a security vulnerability because the
stack could collide with the heap. This change places the stack below
TEXT_START with enough space for 2 MB of stack. This also results in increased
locality of memory which will result in less page faults for small programs.