Some corrections / clarifications for hello-world example (#976)
There are a few typos and confusing text that may cause some confusion
to someone exploring this ground for the first time. It also fixes the
the link to the `examples ` directory at the bottom of the tutorial.
Note that the **terminology** links there are also broken, but
apparently due to how they are bieng interpreted by github.

---------

Co-authored-by: tbro <tbro@users.noreply.github.com>
Co-authored-by: Frank Laub <flaub@risc0.com>