Use PBKDF2 in the password checker example (#910)
This PR changes the password checker example to use PBKDF2 SHA-256 in
the password checker, using the RustCrypto crates.

I had the thought that we actually can simply run PBKDF2 in the guest,
and furthermore could actually prove a PBKDF2 derivation with reasonably
secure iteration counts (e.g. 500,000) with Bonsai.
Now, this would be likely never be worth doing, but it's possible and
showing how easily we can important and use existing crates, which would
be huge pain to implement in another zk system, is a boon to this
exmaple.

On the other hand, this PR brings in Cargo patching and the crypto
accelerated forks.
We may wish to avoid adding that cogantive overhead.
This PR took me 15 minutes to put together and I am happy to throw it
away. :)