cargo-risczero new: add `--guest-name` option
If this is not set, `cargo risczero new` command will prompt the user to enter
a guest name. The guest name parameter is used as the name of the guest package
as well as method ID and ELF constants in the guest code. Therefore, this
parameter is required to be a valid rust identifier. If no value is supplied by
the command line parameter or the prompt, the default value for this parameter
will be "method".