cargo-risczero new: add --guest-name option (#1032)
The guest name parameter is used as the name of the guest package
as well as method ID and ELF constants in the guest code. Therefore,
this
parameter is required to be a valid rust identifier. If this is not set,
the
`cargo risczero new` command will prompt the user to enter a guest name.
If no value is supplied by the command line parameter or the prompt, the
default value for this parameter will be "method".

closes: https://github.com/risc0/scratch/issues/102
closes: #470

---------

Co-authored-by: Frank Laub <flaub@risc0.com>