Add #[inline(never)] to sys_halt and sys_pause
Add #[inline(never)] to sys_halt and sys_pause as a mitigation to
potentially leaking information about program execution through the
final value of the program counter (pc) on halt or pause where there is
more than one location in the program where `sys_halt` is called. As
long as the halt instruction only exists in one place within the
program, the pc will always be the same invariant with input.

Note that this is still not a perfect mitigation, in particular for
privacy of the program itself. Future revisions of the r32im and
recursion circuits will address this issue.