Initial implementation of proof composition (#992)
Proof composition is a new feature of the zkVM allowing verification of
receipts produced by other zkVM guest programs inside of a guest
program. This allow for composition of guest programs, in which the
journal, or other output state, of one guest can be consumed as input to
another.

This PR provides a basic implementation of proof composition. In
particular, it implements composition over the `CompositeReceipt` struct
(formerly `SegmentSeceipts` and `InnerReceipt::Flat`). It does not
implement the recursion predicates needed to produce a single
`SuccintReceipt` for a receipt of execution for a proof composition
(#991).

Other changes:
* This PR clarifies and makes more consistent the semantics of the
`verify` method on `Receipt` and related structs. Semantics are now to
require an exit code of `Halted(0)` or `Paused(0)`, which aligns with
how receipts are using in all examples, and what seems most likely to be
intended by users. It adds the `verify_integrity_with_context` method as
the core method for verifying cryptographic integrity and internal
consistency, allowing use cases that do indeed want to verify receipts
for non-success executions.
* Relatedly, the `prove_guest_errors` options is added to `ProverOpts`
to continue and prove sessions that end with an error status (e.g.
`ExitCode::Fault` or `ExitCode::Halted(1)`). When not set, the prover
will stop after executor on guest error, such that the default is to
only return receipts that of successful execution.

Co-authored-by: Erik Kaneda <erik@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>
Co-authored-by: Matheus Cardoso <45436839+Cardosaum@users.noreply.github.com>