guest code: add `enable-getrandom` feature to toggle getrandom in the guest
Until now, the default behavior was to expose a `getrandom` implementation to
the guest that will call sys_random to generate random numbers from the host.
We did this because the getrandom crate is a dependency of various crates and
without exposing a getrandom function, some crates may fail to compile for the
zkvm. While this is functionality allows more crates to compile, it could
unintentionally allow guest code depend on the host for random numbers.

To avoid a developer unintentionally depending on the guest for random numbers,
change the default behavior to supply a getrandom implementation that panics.
By doing so, developers can be more intentional in allowing or disallowing the
guest code to ask the host to generate random numbers for `getrandom`.