Executor: fix corner case that causes infinite segment split
The executor checks the cycle count for each instruction to ensure that it can
fit within the segment cycle limit. Previously, this was calculated by taking a
measure of the cycle count before the instruction fetch and after execution and
subtracting the two cycle count measurements to determine the cost of that
single instruction. This works for all cases except for the first instruction
of the segment. In this case, the initialization of the segment will incur
several thousand cycles which must be accounted for with in the first
instruction. When checking the first instruction, we must account for the cycle
count accumulated prior to the instruciton fetch.