Executor: simplify logic for instructions that result in large cycle counts (#1186)
The executor checks the cycle count for each instruction to ensure that
it can fit within the segment cycle limit. Previously, this was
calculated by taking a measure of the cycle count before the instruction
fetch and after execution and subtracting the two cycle count
measurements to determine the cost of that single instruction. However,
this approach reveals corner cases. This PR changes the approach to
handling instructions + paging that result in large cycle counts.

If the total pending cycle count exceeds the segment limit and this is
the first instruction, it means that this instruction will never fit
inside the current segment size. Use this instead of the previous
approach to bail on the executor. This a much more simple approach.

Fixes: #1169