fix: Add exponential backoff to bonsai proof polling (#1118)
Currently, this future will essentially spin loop and spam requests to
the service while the proof is waiting to be completed. This doesn't
error for small proofs, but in my testing, this would consistently error
if there was some combination of requests being sent too fast or there
was a proof that took a long time to execute. This caused, on my OS, the
error documented in the issue below as there were too many connections
being opened.

Original issue: https://github.com/risc0/risc0/issues/842

Opening as a draft, as it's opinionated how this is fixed. I can also
add an issue to this if it's helpful, but I figured in this case a hacky
solution might make it more clear what the issue is.

What these changes are is just adding an exponential backoff to this
request, which intuitively makes the most sense to me. Adding a `sleep`
also works, but this is better to minimize latency for small proofs.

---------

Co-authored-by: Matheus <cardosaum@pm.me>
Co-authored-by: Frank Laub <flaub@risc0.com>