Add check for github references in version api docs (#1228)
Working on the docs, I noticed quite a few links to GitHub not using the
correctly pinned version reference (I also contributed to this). I wrote
up a quick script, starting from the `license-check.py` script, to look
for github links in the versioned API docs and ensure they reference the
release branch corresponding to their directory. I've added this rule as
a check in CI.