zkVM: remove SYS_READ_AVAIL and change SYS_READ
This change removes SYS_READ_AVAIL because it's not posix compliant and
introduces some difficulty in cases where the amount of available bytes might
be undefined or non-deterministic with respect to guest execution.

This change removes the SYS_READ_AVAIL support from the host and guest. The
guest syscall libraries used SYS_READ_AVAIL to create a buffer of the correct
size. Due to its removal we're allowing cases where EOF is encountered earlier
and a shorter read than requested is possible through the guest syscall
handlers provided. The underlying handler in the prover has changed to
reflect this reality by removing asserts. This is based on @austinabell's
solution with minor adjustments.

Summary of guest-side syscall handler changes:

Previous implementation relied on the number of bytes returned by
SYS_READ_AVAIL or has been requested. Logic in SYS_READ has been changed to use
the actual amount of bytes read from the syscall. Also, if the guest requests
more bytes than available, it short circuits the requests and returns the
amount of read bytes.

Summary of host changes:

Changed the host to allow allowed the guest to request more bytes than
available. In this case, the host will hit an EOF and return the buffer to the
caller while still allowing last unaligned word to be read.

Co-authored-by: Austin Abell <austinabell8@gmail.com>