Add a basic example of composition (#1274)
This PR adds a basic example of composition. Building upon the proof of
known factors, this example exponentiates a number mod a composite with
known factors. This takes inspiration from RSA encryption, and is kind
of a stripped down (and insecure) version of verifiable encryption such
that the message is provably decrypt-able by some party.

Resolves: #1299

---------

Co-authored-by: Frank Laub <flaub@risc0.com>