Pin cargo-platform to 0.1.5 in zkvm-methods-std (#1324)
An increased minimum required Rust version in the `cargo-platform`
dependency has broken builds of the `release-0.18` branch. As a result,
PRs changing the website (e.g. https://github.com/risc0/risc0/pull/1314)
as stuck since CI requires doc tests to be run against the
`release-0.18` branch.

This PR addresses this issue by updating the `Cargo.lock` file for
`risc0-zkvm-methods-std`, and checking in that file. This matches how
`Cargo.lock` is handled for `risc0-zkvm-methods-guest`, which is the
`no-std` sibling of this package.