zkVM: remove proof of fault
Proof of fault is a mechanism that allows receipt generation for programs that
fail due to incorrect RISC-V execution. We're removing this feature for the
timebeing and plan to implement this in a future circuit revision.

This change removes the fault checker guest code as well as the `fault-proof`
feature flag, and the `bootstrap-fault` command from `cargo xtask`.