zkVM: remove proof of fault (#1341)
Proof of fault is a mechanism that allows receipt generation for
programs that fail due to incorrect RISC-V execution and other
conditions. We're removing this feature for the time being and plan to
implement this in a future circuit revision.

This change removes the fault checker guest code, the `fault-proof`
feature flag, the `bootstrap-fault` command from `cargo xtask`, and
removes `ExitCode::Fault` and marks this as reserved in base.proto.

Fixes: #1307