Updates to Proof System Docs (#1361)
This PR started in response to some broken links around receipt types on
the proof system overview page.

While fixing those links, I went ahead and made some revisions to the
docs to reflect the three layer proving system. This includes adding
some information about our Groth16Receipts, and revising verbiage for
our STARK protocol to reflect the fact that it's one piece in the
picture.

This PR also adds the following terms to the terminology page:
  - deterministic builds
  - Groth16
  - Groth16 Receipt
  - Groth16 Circuit
  - Recursion Circuit
  - RISC-V Circuit

Note: There's currently a [ReceiptClaim] link on the proof system page
which is broken and which this PR does not address. I think this will
work as soon as 0.20 lands; folks should be aware when editing docs that
our proof system docs aren't currently part of the versioned docs system

TODO - images aren't displaying in vercel preview.

---------

Co-authored-by: Tim Zerrell <tim.zerrell@risczero.com>