reproducible builds: update dockerfile to use new toolchain (#1401)
This change downloads a newer version of cargo-risczero and to build
using a newer toolchain version. This solution expects us to publish new
docker images on every toolchain update. While this may create more
burden on us, I think this is much more performant for users because
they would not need to install the toolchain when running docker builds.

---------

Co-authored-by: capossele <angelocapossele@gmail.com>