fix: delete `TempDir` when Session goes out of scope
This solves the problem where the temporary directory used for storing
segments are not deleted after proving. One solution is to express the
directory as `Rc<TempDir>` instead of a `PathBuf` and have `Session` or
`SegmentRef` hold a reference to the `TempDir`. Once the session goes
out of scope, the `TempDir` gets deleted. I've discovered that adding
`TempDir` as a field of `Session` doesn't work because `TempDir` does
not implement Serialize or Deserialize while Session and `SegmentRef`
implement these traits.

This solution adds the `PathBuf` of the `TempDir` to `Session` and
implements a `Drop` trait that removes this directory when Session goes
out of scope. I've also noticed that in the case where the user defines
the segment path, we're using that path to store the segments. This
changes this execution path by creating a `TempDir` inside the user-
defined directory so that it does not accidentally delete the user's
directory upon termination.