Clean up segment path by using `Rc<TempDir>`
At this time, the zkVM fails to clean up segment information. Fix this
by reference-counting each segment file if it uses a tempdir. We will
let the user manage the lifetime of the segment if they designate a
specific path to store segments. Some of this is based off of Austin's
work.

Co-Authored-by: Austin Abell <austinabell8@gmail.com>