CI: update cargo-install GH action to 2.0 (#1431)
By updating to 2.0, this action will use the `--locked` option by
default and will use the semver listed in the yml files. Also, pin
`clap` to 4.4 and use an earlier version of `cargo-rdme`. `cargo-rdme`
will be updated once we update host-side toolchains to 1.74.0 or
higher.

The 0.18 and 0.19 doc tests have been updated to pin to a specific version
because those refer to versions of our software that doesn't have a
`Cargo.lock` file. Going forward, we will add `.lock` files.

Future PR will remove the doc test for 0.18 and 0.19.