Remove reference to relay from Bonsai overview docs (#1470)
Following up on https://github.com/risc0/risc0/pull/1469 this PR removes
references to the relay from the Bonsai overview doc.

In PR(s) following this one, we'll have a more complete refactoring and
rewriting of the docs to reflect the up to date patterns for the zkVM
and Bonsai.