sidebar reorganization (#1468)
This PR mostly just moves content around in order to make the conceptual
flow clearer.
Rather than looking at the diff in GitHub, I suggest reviewing by
opening the deployment, imagining yourself as a new user, and trying it
out.

The high-level thinking around the conceptual flow for users is
described in this diagram:
<img width="1046" alt="image"
src="https://github.com/risc0/risc0/assets/15272444/08f89eee-adc6-48da-acf2-b905475f9ce2">

- [ ] This PR moves the `bonsai/quickstart.md` doc to
`generating-proofs/remote-proving.md`. If we use this organization, we
need to set up a re-direct so that the QR codes and outstanding links
that point to `bonsai/quickstart` don't 404.
- [ ] This PR deletes the `bonsai/overview.md` doc. We need to set up a
re-direct so that QR codes and outstanding links that point to `bonsai/`
don't 404. The redirect should point to
`generating-proofs/remote-proving.md`

<s>This PR also drops the slug on the `zkvm_overview` doc, which means
links to `/zkvm/` will 404. If we want to drop this slug, we should set
up a re-direct from `/zkvm/` to `/zkvm/zkvm_overview.md` to address
this. </s> resolved by keeping the slug

---------

Co-authored-by: Victor Graf <victor@risczero.com>
Co-authored-by: Frank Laub <frank@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>