Remove optimizations from build scripts (#1447)
Throughout our repo we apply optimizations to Cargo build scripts (i.e.
binaries built to run as part of the build process).
This was introduced to make calculation of control IDs faster, as this
was the bottleneck for the build process.

Now that control IDs are not calculated on a per-guest basis, it is
unclear if optimizing the build scripts is still beneficial to build
time.
Opening this small PR to see what the build times are in CI across the
various examples and tests without this setting.

Co-authored-by: Erik Kaneda <erik@risczero.com>