Revert "risc0-build: move the env var check for RISC0_SKIP_BUILD (#1419)"
This reverts commit c68575c3bea0f2bcf275dbc60d99fc59dd73dbdd and moves
the logic to skip build inside docker and non-docker build functions.

Moving the env var to the embed_methods results in skipping the
embedding of methods in addition to skipping the build. By doing this,
RISC0_SKIP_BUILD will only skip the build and continue to embed the
methods.rs file containing the ELF and image ID.