changing `resources` to `education hub` (#1505)
In the status quo, it feels awkward/jarring that when you click on
"resources," you land directly on the study club page. More broadly, our
"resources" section feels a bit weak, and I think we have a lot of
educational assets to show off.

This PR rebrands the "resources" section as an "Education Hub" and adds
a landing page to orient users to the contents of that section.
Previously, clicking "resources" pointed users to the study club page in
a way that felt rather jarring to me.

On a follow-up PR, I think we should probably pop the proof system docs
out of this section.