CI: add cargo-semver-checks
This adds an action to check an in-coming PR against the target to point
out any changes in the crates we publish that might break semver. This
implementation lists a few packages that we exclude from the check.
Among the excluded are packages that do not need to be published, will
be moved, or ones that cargo semverchecks cannot locate correctly due to
a possible bug (`risc0-zkvm-platform`).

This change is designed for us to get used to seeing what sort of
changes break semver and is intended to not block PR's from getting at
this time.