DRAFT: Implement syscall for receipt verification
This commit implements a new syscall allowing the guest to verify a
specific receipt, rather than simply verifying that *some* receipt is
known which satisfies the claim.

This commit implements the requisite syscall but omits the crucial step
of recursively invoking the Rust verifier when the host claims that the
provided receipt is invalid. This omission is intentional, since the
Risc0 verifier currently does not run on Risc0.