Build rust crates that bind to C code (#1610)
This PR is the first step in enabling crates that are bindings to c
code. This PR changes cargo-risczero to download a prebuilt c toolchain,
and sets appropriate flags on risc0-build to invoke the c compiler when
building guest crates. This solution is able to compile the blst crate
so I've added a test case that exercises this crate. Note: while this
exercises a signature verification example, further testing is needed to
conclude that BLST C code is working as expected by using utilities such as
`cargo risczero test`.

---------

Co-authored-by: Frank Laub <flaub@risc0.com>