Validate Poseidon `hash_pair` input (#1372)
Our Poseidon/Poseidon2 `hash_pair` functionality is intended to work
with Digests generated by Poseidon hashing. However, prior to this PR,
it does not perform checking to ensure that its input Digests are in
this format. This means a developer unaware of this nuance might
accidentally write code that expects these `hash_pair` functions to work
as hash functions on general Digests with values outside those possible
from Poseidon hashing, which could create a vulnerability depending on
the use case.

This PR adds an `assert` that validates all inputs are in the required
format by adding the ability to test that words in a Digest are elements
of the underlying field in reduced form. Thus, any developer misusing
`hash_pair` in the above way would hit this `assert` instead of running
to completion and producing a "hash" that does not have all the security
properties such a developer might expect.

---------

Co-authored-by: Victor Graf <victor@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>