Make list of allowed ZKRs explicit in bootstrap (#1614)
Currently, the `test_recursion_circuit` ZKR is included in the allowed
list even though it is only for use in tests. The allowed set of ZKRs
should only include operations that are meant to be applied to succinct
and segment receipts. This PR adds an explicit list of the ZKRs to be
included in the allowed set. This means that the test ZKR(s), and
accelerators in the future, can be included in the zip, without being
automatically included in the allowed set.

---------

Co-authored-by: Frank Laub <frank@risczero.com>