Change Elem from bytemuck::Pod to NoUninit + CheckedBitPattern (#1373)
Building on https://github.com/risc0/risc0/pull/1372 , this PR adjusts
the type system to try to systematically prevent misuse casting into the
`Elem` type. Core to this is that the `Elem` type actually does not
satisfy the `bytemuck::Pod` safety contract as there are some bit
patterns which are not valid. In particular, any `u32` greater than `P`
is not valid as field element, and the arithmetic code does not handle
these values.

As a start, this PR change `Elem` from being marked as `Pod` to
`NoUninit` and `CheckedBitPattern` together these methods allow for
partial functionality from `bytemuck`. In particular, `NoUninit` can be
the source type for `bytemuck::cast_slice` and `CheckedBitPattern` can
be the target for `bytemuck::checked::cast_slice`. Note that the latter
introduces a runtime check that the `u32` input is less than `P`.

In the verifiers, only these safe methods should be used. In the prover,
there may be times when we want to cast from bytes or words to elements
without any checks, to maximize performance. It is my aim to preserve
this through methods marked explicitly as unchecked.

---------

Co-authored-by: Tim Zerrell <tim.zerrell@risczero.com>
Co-authored-by: Frank Laub <frank@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>