Rename get_claim methods to claim to follow C-GETTER convention (#1716)
A quick PR for renaming the `get_claim` method to `claim`, following the
`C-GETTER` convention recommended by Rust API guidelines.

https://rust-lang.github.io/api-guidelines/naming.html#getter-names-follow-rust-convention-c-getter

I'm not very attached to getting this merged, but figured it is an easy
enough PR to throw together and only possible now, before 1.0. This
change doesn't break any examples, which I take as a sign that it will
not break most user code. If folks disagree that we should make this
change, I'll go ahead and close it.

Co-authored-by: Frank Laub <flaub@risc0.com>